Refined Ghost Movement...Added Ghost Spell
I refined the Ghost Movement to be using Jakes movement update
functions and direction pattern.  Previously, it was using my makeshift
movement logic.

I also created a crappy version of a laser implement as triangles for
now to get a working version of a spell.  This will be fired by the
ghost object.

I also added a new function to map.cpp which allows you to fire a spell
on the map based on direction and not calculated from the mouse
pointer.  I want to eventually have the laser fire towards the player
so I figured the ghost could calculate that in its class and not the
map.  May need to be changed